refactor(skills): relocate eval dashboard into arc-evaluating/dashboard
Move eval-dashboard.js and eval-dashboard-ui.html from scripts/ to
skills/arc-evaluating/dashboard/. Update require paths in dashboard.js
to use ../../../scripts/lib/. Update cli.js to reference new location.
Relocate dashboard test alongside its module and expand jest.config.js
testMatch to discover skill-local __tests__ directories.
